feat(web): guided ingestion in settings + lean Connect launchpad
Unify Maple's "first telemetry" experience across three surfaces that
previously didn't share code. The guided flow (framework picker, install/
instrument/Claude Code tabs, live "watching for your first trace" status)
lived only on the dashboard setup checklist, while the Connect popover and
Settings → Ingestion each re-implemented their own flat copy-field stacks
(maskKey + CopyableField were triplicated).

- New shared module apps/web/src/components/ingest/: CopyableField + maskKey,
  useIngestConnection (live service-overview poll, filters demo- services) +
  sendTestEvent, ConnectCredentials, GuidedSetup, and connection-status
  (status pill / waiting strip / persistent settings panel).
- Settings → Ingestion: new "Send your first telemetry" guided card on top
  (framework picker + tabbed snippets with credentials interpolated, live
  waiting↔connected status) above the existing endpoint/keys/mappings.
- Connect popover: lean launchpad — live "Connected · N services" status pill,
  compact credentials, Claude Code hero path, "Open setup guide →" deep-link
  into Settings → Ingestion.
- Dashboard setup checklist refactored onto the shared module; collapse/
  dismiss/celebration behavior preserved.
